Implement video handling and restructure the code
Implement 'video_to_images' function which reads a video from input
video filename and saves each frame as an individual jpg. Put the actual
image processing in a separate function which takes a cv2 image as input
and saves processed full and crop versions of the image in suitable sub
directories. Implement 'image_loop' which loops over all images in a
directory.